requirements.txt
aiogram
asyncpg
python-dotenv
redis
geopy